continue splitting up generators
