+++ src/generators/members/index.ts
@@ -1,2 +1,3 @@
 export * from "./server-constructor.js";
+export * from "./server-methods.js";
 export * from "./server-properties.js";
