+++ src/generators/members/server-properties.ts
@@ -1,15 +1,19 @@
 import ts from "typescript";
+import * as models from "../../models/index.js";
+import { toCamel, toPascal } from "../../utils/name.js";
 import { CodeGeneratorBase } from "../code-generator-base.js";
 
 export class ServerPropertiesCodeGenerator extends CodeGeneratorBase {
   public *getStatements() {
     yield* this.generateRouterPropertyStatements();
+    yield* this.generateOperationHandlersPropertiesStatements();
+    yield* this.generateAuthorizationHandlersPropertiesStatements();
   }
 
   /**
    * the router property
    */
-  protected *generateRouterPropertyStatements() {
+  private *generateRouterPropertyStatements() {
     const { factory: f } = this;
 
     const identityFunctionExpression = f.createArrowFunction(
@@ -56,6 +60,85 @@ export class ServerPropertiesCodeGenerator extends CodeGeneratorBase {
       undefined,
       undefined,
       newRouterExpression,
+    );
+  }
+
+  /**
+   * operation handler properties that may contain operation handlers
+   */
+  private *generateOperationHandlersPropertiesStatements() {
+    const { factory: f } = this;
+
+    for (const pathModel of this.apiModel.paths) {
+      for (const operationModel of pathModel.operations) {
+        yield* this.generateOperationHandlersPropertyStatements(
+          pathModel,
+          operationModel,
+        );
+      }
+    }
+  }
+
+  /**
+   * a single property yo hold the operation handler
+   * @param pathModel
+   * @param operationModel
+   */
+  private *generateOperationHandlersPropertyStatements(
+    pathModel: models.Path,
+    operationModel: models.Operation,
+  ) {
+    const { factory: f } = this;
+
+    const operationHandlerPropertyName = toCamel(
+      "handle",
+      operationModel.name,
+      "operation",
+    );
+    const operationHandlerTypeName = toPascal(
+      operationModel.name,
+      "operation",
+      "handler",
+    );
+
+    yield f.createPropertyDeclaration(
+      [f.createToken(ts.SyntaxKind.PrivateKeyword)],
+      f.createIdentifier(operationHandlerPropertyName),
+      f.createToken(ts.SyntaxKind.QuestionToken),
+      f.createTypeReferenceNode(operationHandlerTypeName),
+      undefined,
+    );
+  }
+
+  private *generateAuthorizationHandlersPropertiesStatements() {
+    for (const authorizationModel of this.apiModel.authorizations) {
+      yield* this.generateAuthorizationHandlersPropertyStatements(
+        authorizationModel,
+      );
+    }
+  }
+
+  private *generateAuthorizationHandlersPropertyStatements(
+    authorizationModel: models.Authorization,
+  ) {
+    const { factory: f } = this;
+
+    const authorizationHandlerName = toCamel(
+      "handle",
+      authorizationModel.name,
+      "authorization",
+    );
+
+    yield f.createPropertyDeclaration(
+      [f.createToken(ts.SyntaxKind.PrivateKeyword)],
+      f.createIdentifier(authorizationHandlerName),
+      f.createToken(ts.SyntaxKind.QuestionToken),
+      f.createFunctionTypeNode(
+        undefined,
+        [],
+        f.createKeywordTypeNode(ts.SyntaxKind.VoidKeyword),
+      ),
+      undefined,
     );
   }
 }
